Update committee quorum (#7)
Increase base and executive quorum requirements by one in order to require a majority. Currently we only require 50%, which is not sufficient to reasonably be considered a majority.

Add note specifying that nonvoting Committee positions do not count towards quorum, to make it explicit that IPP and any future nonvoting positions being present does not allow a sub-quorum of voting members to decide matters.